Split UrlRequest_Windows.cpp into separate files for UWP and Win32
